requirements.txt
keyboard~=0.13.5
pyautogui~=0.9.54
pydirectinput~=1.0.4
pyscreeze
Pillow